# fix(node): Convert debugging code to callbacks to fix memory leak in `LocalVariables` integration