--- a/packages/node/test/integrations/localvariables.test.ts
+++ b/packages/node/test/integrations/localvariables.test.ts
@@ -4,7 +4,7 @@ import type { LRUMap } from 'lru_map';
 
 import { defaultStackParser } from '../../src';
 import type { DebugSession, FrameVariables } from '../../src/integrations/localvariables';
-import { LocalVariables } from '../../src/integrations/localvariables';
+import { createCallbackList, LocalVariables } from '../../src/integrations/localvariables';
 import { getDefaultNodeClientOptions } from '../../test/helper/node-client-options';
 
 interface ThrowOn {
@@ -44,7 +44,7 @@ class MockDebugSession implements DebugSession {
 }
 
 interface LocalVariablesPrivate {
-  _cachedFrames: LRUMap<string, Promise<FrameVariables[]>>;
+  _cachedFrames: LRUMap<string, FrameVariables[]>;
   _setup(addGlobalEventProcessor: (callback: EventProcessor) => void, clientOptions: ClientOptions): void;
 }
 
@@ -108,6 +108,43 @@ const exceptionEvent = {
   },
 };
 
+const exceptionEvent100Frames = {
+  method: 'Debugger.paused',
+  params: {
+    reason: 'exception',
+    data: {
+      description:
+        'Error: Some  error\n' +
+        '    at two (/dist/javascript/src/main.js:23:9)\n' +
+        '    at one (/dist/javascript/src/main.js:19:3)\n' +
+        '    at Timeout._onTimeout (/dist/javascript/src/main.js:40:5)\n' +
+        '    at listOnTimeout (node:internal/timers:559:17)\n' +
+        '    at process.processTimers (node:internal/timers:502:7)',
+    },
+    callFrames: new Array(100).fill({
+      callFrameId: '-6224981551105448869.1.0',
+      functionName: 'two',
+      location: { scriptId: '134', lineNumber: 22 },
+      url: '',
+      scopeChain: [
+        {
+          type: 'local',
+          object: {
+            type: 'object',
+            className: 'Object',
+            objectId: '-6224981551105448869.1.2',
+          },
+          name: 'two',
+        },
+      ],
+      this: {
+        type: 'object',
+        className: 'global',
+      },
+    }),
+  },
+};
+
 describe('LocalVariables', () => {
   it('Adds local variables to stack frames', async () => {
     expect.assertions(7);
@@ -134,15 +171,15 @@ describe('LocalVariables', () => {
 
     expect((localVariables as unknown as LocalVariablesPrivate)._cachedFrames.size).toBe(1);
 
-    let frames: Promise<FrameVariables[]> | undefined;
+    let frames: FrameVariables[] | undefined;
 
-    (localVariables as unknown as LocalVariablesPrivate)._cachedFrames.forEach(promise => {
-      frames = promise;
+    (localVariables as unknown as LocalVariablesPrivate)._cachedFrames.forEach(f => {
+      frames = f;
     });
 
     expect(frames).toBeDefined();
 
-    const vars = await (frames as Promise<FrameVariables[]>);
+    const vars = frames as FrameVariables[];
 
     expect(vars).toEqual([
       { function: 'two', vars: { name: 'tim' } },
@@ -206,6 +243,41 @@ describe('LocalVariables', () => {
     expect(event?.exception?.values?.[0].stacktrace?.frames?.[4]?.vars).toEqual({ name: 'tim' });
 
     expect((localVariables as unknown as LocalVariablesPrivate)._cachedFrames.size).toBe(0);
+  });
+
+  it('Only considers the first 5 frames', async () => {
+    expect.assertions(4);
+
+    const session = new MockDebugSession({});
+    const localVariables = new LocalVariables({}, session);
+    const options = getDefaultNodeClientOptions({
+      stackParser: defaultStackParser,
+      includeLocalVariables: true,
+    });
+
+    let eventProcessor: EventProcessor | undefined;
+
+    (localVariables as unknown as LocalVariablesPrivate)._setup(callback => {
+      eventProcessor = callback;
+    }, options);
+
+    expect(eventProcessor).toBeDefined();
+
+    await session.runPause(exceptionEvent100Frames);
+
+    expect((localVariables as unknown as LocalVariablesPrivate)._cachedFrames.size).toBe(1);
+
+    let frames: FrameVariables[] | undefined;
+
+    (localVariables as unknown as LocalVariablesPrivate)._cachedFrames.forEach(f => {
+      frames = f;
+    });
+
+    expect(frames).toBeDefined();
+
+    const vars = frames as FrameVariables[];
+
+    expect(vars.length).toEqual(5);
   });
 
   it('Should not lookup variables for non-exception reasons', async () => {
@@ -287,5 +359,73 @@ describe('LocalVariables', () => {
     await session.runPause(exceptionEvent);
 
     expect((localVariables as unknown as LocalVariablesPrivate)._cachedFrames.size).toBe(1);
+  });
+
+  describe('createCallbackList', () => {
+    it('Should call callbacks in reverse order', done => {
+      const log: number[] = [];
+
+      const { add, next } = createCallbackList<number>(n => {
+        expect(log).toEqual([5, 4, 3, 2, 1]);
+        expect(n).toBe(15);
+        done();
+      });
+
+      add(n => {
+        log.push(1);
+        next(n + 1);
+      });
+
+      add(n => {
+        log.push(2);
+        next(n + 1);
+      });
+
+      add(n => {
+        log.push(3);
+        next(n + 1);
+      });
+
+      add(n => {
+        log.push(4);
+        next(n + 1);
+      });
+
+      add(n => {
+        log.push(5);
+        next(n + 11);
+      });
+
+      next(0);
+    });
+
+    it('only calls complete once even if multiple next', done => {
+      const { add, next } = createCallbackList<number>(n => {
+        expect(n).toBe(1);
+        done();
+      });
+
+      add(n => {
+        next(n + 1);
+        // We dont actually do this in our code...
+        next(n + 1);
+      });
+
+      next(0);
+    });
+
+    it('calls completed if added closure throws', done => {
+      const { add, next } = createCallbackList<number>(n => {
+        expect(n).toBe(10);
+        done();
+      });
+
+      add(n => {
+        throw new Error('test');
+        next(n + 1);
+      });
+
+      next(10);
+    });
   });
 });