Closes #7230 

This PR converts the `LocalVariables` debugger code to callbacks so that we can call `Debugger.resume` synchronously from the `pause` callback which fixes the memory leak.

It's worth noting that now there is no usage of async, it will not be possible to use async import to improve this:
https://github.com/getsentry/sentry-javascript/blob/0743e988d6e160a530b07cb80b9ec4cb17ba9b39/packages/node/src/integrations/localvariables.ts#L32-L35

I created a basic Express server to test these changes and load tested an endpoint with a caught/handled error to see if any memory is leaked. Without the async code, performance is also much improved:

|  | Before | After |
| --- | --- | --- |
| Requests/sec | 40 | 220 |
| Av. Latency | 90ms | 40ms |
| Peak Memory | 570 MB  | 130 MB  |